# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gym Me BackEnd Thesis",
        "title": "Gym Me Backend Thesis",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "dasdasdas",
            "url": "asdasdas",
            "email": "tes@gmail.com"
        },
        "license": {
            "name": "Gymme"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "paths": {
...
        "/api/calendar": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create New Calendar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Create New Calendar",
                "parameters": [
                    {
                        "description": "Insert Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MenuPayloads.CalendarInsertPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StandarAPIResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Calendar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Update Calendar",
                "parameters": [
                    {
                        "description": "Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MenuPayloads.CalendarUpdatePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    }
                }
            }
        },
...
        "/api/calendar/delete/{calendar_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Calendar by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Delete Calendar by Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "calendar_id",
                        "name": "calendar_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StandarAPIResponses"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "MenuPayloads.CalendarInsertPayload": {
            "type": "object",
            "properties": {
                "calendar_date": {
                    "type": "string"
                },
                "calendar_name": {
                    "type": "string"
                },
                "calendar_time_from": {
                    "type": "string"
                },
                "calendar_time_to": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "MenuPayloads.CalendarUpdatePayload": {
            "type": "object",
            "properties": {
                "calendar_date": {
                    "type": "string"
                },
                "calendar_id": {
                    "type": "integer"
                },
                "calendar_name": {
                    "type": "string"
                },
                "calendar_time_from": {
                    "type": "string"
                },
                "calendar_time_to": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "responses.ErrorResponses": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.StandarAPIResponses": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}